+++ app/cart/models/cart.js
@@ -1,55 +1,10 @@
 const mongoose = require("mongoose");
-const schema = new mongoose.Schema({
-  _id: {
-    type: String,
-    required: true,
-  },
-  email: {
-    type: String,
-    required: true,
-  },
-  username: {
-    type: String,
-    required: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  name: {
-    firstname: {
-      type: String,
-    },
-    lastname: {
-      type: String,
-    },
-  },
-  address: {
-    city: {
-      type: String,
-    },
-    street: {
-      type: String,
-    },
-    number: {
-      type: Number,
-    },
-    zipcode: {
-      type: Number,
-    },
-    geolocation: {
-      lat: {
-        type: Number,
-      },
-      long: {
-        type: Number,
-      },
-    },
-  },
-  phone: {
-    type: Number,
-  },
+const Schema = mongoose.Schema;
+const cartSchema = new Schema({
+  _id: String,
+  userId: Number,
+  date: String,
+  time: String,
+  products: Array,
 });
-
-const User = mongoose.model("User", schema);
-module.exports.User = User;
+module.exports = mongoose.model("cart", cartSchema, "carts");
