+++ app/user/controller/userControllers.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-const productModel = require("../models/user");
+const User = require("../models/user");
 
 class UserController {
   constructor() {}
