+++ app/user/models/user.js
@@ -1,12 +1,55 @@
 const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-const productSchema = new Schema({
-  _id: String,
-  productTitle: String,
-  productPrice: Number,
-  productDescription: String,
-  productCategory: String,
-  productImageName: String,
-  productImagePath: String,
+const schema = new mongoose.Schema({
+  _id: {
+    type: String,
+    required: true,
+  },
+  email: {
+    type: String,
+    required: true,
+  },
+  username: {
+    type: String,
+    required: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+  name: {
+    firstname: {
+      type: String,
+    },
+    lastname: {
+      type: String,
+    },
+  },
+  address: {
+    city: {
+      type: String,
+    },
+    street: {
+      type: String,
+    },
+    number: {
+      type: Number,
+    },
+    zipcode: {
+      type: Number,
+    },
+    geolocation: {
+      lat: {
+        type: Number,
+      },
+      long: {
+        type: Number,
+      },
+    },
+  },
+  phone: {
+    type: Number,
+  },
 });
-module.exports = mongoose.model("product", productSchema, "products");
+
+const User = mongoose.model("User", schema);
+module.exports.User = User;
